GH-397: [fix & refactor] remove auto-scroll to post, make post interactive in feed list, move types to corresponding types files, present error with toast
