+++ Frontend/app/(app)/_layout.tsx
@@ -29,7 +29,6 @@ export default function AppLayout() {
         options={{
           title: "",
           presentation: "formSheet",
-          // sheetAllowedDetents: "fitToContents",
           sheetAllowedDetents: [0.4, 1],
           contentStyle: {
             backgroundColor: "transparent",
