+++ Frontend/app/(app)/leagues/[id]/index.tsx
@@ -121,12 +121,8 @@ export default function LeagueScreen() {
 function LeagueContent() {
   const params = useLocalSearchParams<{
     tab?: string;
-    postId?: string | string[];
   }>();
   const initialTab: LeagueTab = resolveLeagueTab(params.tab);
-  const targetPostId = Array.isArray(params.postId)
-    ? params.postId[0]
-    : params.postId;
   const [tab, setTab] = useState<LeagueTab>(initialTab);
   const router = useRouter();
   const log = createScopedLog("League Page");
@@ -279,7 +275,6 @@ function LeagueContent() {
                 }
                 onDeletePost={handleDeletePost}
                 canDelete={isOwner}
-                targetPostId={targetPostId}
               />
             )}
