+++ Frontend/app/(app)/teams/[id]/index.tsx
@@ -120,12 +120,8 @@ export default function Team() {
 function TeamContent() {
   const params = useLocalSearchParams<{
     tab?: string;
-    postId?: string | string[];
   }>();
   const initialTab: TeamTab = parseTeamTab(params.tab);
-  const targetPostId = Array.isArray(params.postId)
-    ? params.postId[0]
-    : params.postId;
   const [tab, setTab] = useState<TeamTab>(initialTab);
   const router = useRouter();
   const log = createScopedLog("Team Page");
@@ -305,7 +301,6 @@ function TeamContent() {
                 }
                 onDeletePost={handleDeletePost}
                 canDelete={canManage}
-                targetPostId={targetPostId}
               />
             )}
             {tab === "matches" && (
