+++ Frontend/components/board/post.tsx
@@ -11,7 +11,7 @@ import { useActionSheet } from "@expo/react-native-action-sheet";
 import ContextMenu from "react-native-context-menu-view";
 import { Card } from "@/components/ui/card";
 import { IconSymbol } from "@/components/ui/icon-symbol";
-import { BoardPost } from "@/components/board/board-types";
+import { BoardPost } from "@/types/board";
 import { isRunningInExpoGo } from "@/utils/runtime";
 import TimeAgo from "javascript-time-ago";
 
@@ -21,6 +21,7 @@ interface PostProps {
   spaceLogo: ImageSource;
   onDelete?: (postId: string) => void;
   canDelete?: boolean;
+  isInteractive?: boolean;
   onPress?: () => void;
 }
 
@@ -30,6 +31,7 @@ export function Post({
   spaceLogo,
   onDelete,
   canDelete = false,
+  isInteractive = false,
   onPress,
 }: Readonly<PostProps>) {
   const { showActionSheetWithOptions } = useActionSheet();
@@ -63,7 +65,7 @@ export function Post({
   };
 
   const cardContent = (
-    <Card isInteractive={isRunningInExpoGo}>
+    <Card isInteractive={isInteractive || isRunningInExpoGo}>
       <View style={styles.container}>
         <View style={styles.content}>
           <View style={styles.header}>
