+++ Frontend/types/leagues.ts
@@ -1,3 +1,25 @@
+export type LeagueSummaryResponse = {
+  id: string;
+  name: string;
+  sport: string;
+  slug: string;
+  logoUrl?: string | null;
+  region?: string | null;
+  level?: string | null;
+  privacy?: string | null;
+  seasonCount?: number | null;
+  createdAt: string;
+  updatedAt: string;
+};
+
+export type LeagueListResponse = {
+  items: LeagueSummaryResponse[];
+  totalElements: number;
+  page: number;
+  size: number;
+  hasNext: boolean;
+};
+
 export type LeagueInviteResponse = {
   id: string;
   leagueId: string;
